updateref

+++ server/test/services/applicant/predicate/JsonPathPredicateGeneratorTest.java
@@ -100,7 +100,7 @@ public class JsonPathPredicateGeneratorTest {
             dateQuestion.getId(),
             Scalar.DATE,
             Operator.AGE_OLDER_THAN,
-            PredicateValue.of((float) 18.5));
+            PredicateValue.of((double) 18.5));
 
     JsonPathPredicate predicate =
         JsonPathPredicate.create("$.applicant.applicant_birth_date[?(1309478400000 >= @.date)]");
